fixed small but in routers/user_skills.py
The /user/skills/ POST endpoint did not throw an error when in case the provided skill was invalid - instead it gave 'items': []. Solved by adding [0][0] in line 167 of user_skills.py. I assume the COLLECT sttatement in the cypher query is responsible for this behaviour, giving an empty items list.